backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
PyMuPDF==1.23.21
sentence-transformers==2.3.1
faiss-cpu==1.7.4
groq>=1.0.0
python-dotenv==1.0.1
pydantic==2.5.3
numpy==1.26.3
